refactor: /validate api authController -> userController 로 변경
